Add commands isolated to channel

+++ src/webui/components/panel/commands/actions/DeleteCommand.tsx
@@ -1,12 +1,21 @@
 import { prisma } from "services/db";
 import { TSingleUiCommand } from "webui/types";
 import { SingleCommand } from "../SingleCommand";
+import { Context } from "elysia";
+import { getChannelFromCtx } from "webui/helpers";
+import { getUniqueName } from "services/commands";
 
-export const DeleteCommand = async (props: TSingleUiCommand) => {
+export const DeleteCommand = async (props: TSingleUiCommand, ctx: Context) => {
+  const channel = await getChannelFromCtx(ctx);
+
+  if(!channel) {
+    return <SingleCommand {...props} />;
+  }
+
   try {
     await prisma.commands.delete({
       where: {
-        name: props.name,
+        uniqueName: getUniqueName(props.name, channel.name)
       },
     });
